+++ cpp/src/turn.cpp
@@ -1,6 +1,6 @@
 #include "../include/turn.h"
 
-Turn::Turn(double score_, Coordinate currentLocation_, Move change_){
+Turn::Turn(int score_, Coordinate currentLocation_, Move change_){
     score = score_;
     currentLocation = currentLocation_;
     change = change_;
